Rework spec to match query collaboration sequence
Removing the interaction with birthday because it's not a relevant part
of showing the testing of this query collaboration.

Without the birthday/age code my dumb Pirates of Penzance joke is no
longer relevant. Making our Test Person setup the same across all tests
so as to avoid causing confusion with irrelevant changes.